Optimize 25 Cython problems that benchmarked below 3x speedup
Major improvements (>10x gain):
- salsa20_quarter: 1.76x → 227x (removed redundant & MASK32 on unsigned int, added nogil)
- pascal_triangle_row: 1.52x → 155x (O(n) binomial recurrence vs O(n²) loop)
- classify_transitions: 2.20x → 131x (typed ULL constants, nogil)
- rpn_eval: 2.31x → 117x (stack-allocated C array vs cdef class dispatch)
- suffix_array_lcp: 2.48x → 53x (Manber-Myers prefix-doubling algorithm)
- interval_overlap_count: 1.59x → 43x (C arrays + qsort, in-place filtering)
- two_sum_all_pairs: 1.63x → 33x (C frequency array vs Python dict)
- mobius_sieve: 2.43x → 15x (fused inner loops, C while loops, nogil)

Moderate improvements:
- patience_sort: 2.93x → 6.2x (C-level min-heap + C piles)
- fizzbuzz: 2.34x → 4.6x (custom uint-to-str, PyList_SET_ITEM)
- bisection_batch: 2.89x → 3.0x (added nogil)
- prefix_max: 2.48x → 3.1x (cpython.array memoryview + bulk tolist)
- linked_list_sum: 2.36x → 2.8x (cdef C-level traversal methods)
- typedef_callback_sort: 1.57x → 2.6x (O(n) reversal for 2nd sort)

NumPy problems — removed unnecessary .astype(np.float64) copies across all 9
files (standard_normal/random/linspace already return float64). Additional
optimizations: softmax buffers exp() results, histogram uses reciprocal
multiplication, batch_norm pre-computes fused scale constant.

+++ cnake_charmer/cy/numerical/numpy_typed_interp.pyx
@@ -36,23 +36,15 @@ def numpy_typed_interp(int n):
     """Interpolate n query points and return sum."""
     rng = np.random.RandomState(42)
     cdef int num_knots = 1000
-    cdef cnp.ndarray[cnp.float64_t, ndim=1] xp_arr = (
-        np.linspace(0.0, 1.0, num_knots).astype(
-            np.float64
-        )
-    )
+    cdef cnp.ndarray[cnp.float64_t, ndim=1] xp_arr = np.linspace(0.0, 1.0, num_knots)
     cdef cnp.float64_t[::1] xp = xp_arr
 
-    cdef cnp.ndarray[cnp.float64_t, ndim=1] fp_arr = (
-        np.cumsum(
-            rng.standard_normal(num_knots)
-        ).astype(np.float64)
+    cdef cnp.ndarray[cnp.float64_t, ndim=1] fp_arr = np.cumsum(
+        rng.standard_normal(num_knots)
     )
     cdef cnp.float64_t[::1] fp = fp_arr
 
-    cdef cnp.ndarray[cnp.float64_t, ndim=1] xq_arr = (
-        rng.random(n).astype(np.float64)
-    )
+    cdef cnp.ndarray[cnp.float64_t, ndim=1] xq_arr = rng.random(n)
     cdef cnp.float64_t[::1] xq = xq_arr
 
     cdef cnp.float64_t total = 0.0
